Emacs startup
We setup a literate configuration using babel.

We use straight.el to manage package installation, but we rely
on use-package for configuration. The main reason is to get out of
the misery that using a newer org-mode than what comes with Emacs is
with package.el

We also install restart-emacs in order to be able to quickly restart
Emacs while playing with the configuration.

Last we keep our configuration clean by routing all files generated
by packages to better places.

We also introduce a emacs script for launching emacs. This will become
more useful when we'll start to compile our own emacs distributions.